Refactor AudioThread to prevent self-stop in audio thread
- Updated AudioThread::Stop() to disallow stopping from within the audio thread to avoid unsafe state transitions.
- Added error logging for self-stop attempts.

Change CircularBuffer template parameter from unsigned int to u32

- Modified CircularBuffer::GetMultiple() to use u32 for the count template parameter.

Change compilation time type in CompilerCache

- Changed m_CompilationTimeMs from f64 to f32 in CompilerCache for consistency with other float types.

Enhance ArrayNodes with output triggering method

- Introduced TriggerOutput() method to streamline output handling in ArrayNodes.
- Refactored ProcessTrigger() to utilize TriggerOutput() for cleaner code.

Add profiling to Random node initialization

- Added OLO_PROFILE_FUNCTION() to Random node's Init() method for performance tracking.

Update SoundGraphCache to use configuration struct

- Refactored SoundGraphCache constructor to accept SoundGraphCacheConfig for better configurability.
- Removed hardcoded cache configuration constants.

Improve SoundGraphFactory to use contains() method

- Updated Factory::Contains() to use the modern contains() method for better readability.

Fix SoundGraphPlayer to ensure thread safety during updates

- Held mutex during updates to prevent use-after-free issues.
- Removed unnecessary copying of source references before updates.

Standardize volume and pitch types in SoundGraphSound

- Changed volume and pitch member variables to f32 for consistency.
- Updated related methods to reflect the change in type.

Refactor WaveSource to improve instance function handling

- Introduced SetupInstanceTrampoline() to encapsulate the setup of function pointers for refill callbacks.
- Simplified assignment and copy operations for RefillCallback.

+++ OloEngine/src/OloEngine/Audio/SoundGraph/CompilerCache.h
@@ -35,7 +35,7 @@ namespace OloEngine::Audio::SoundGraph
         std::string m_ErrorMessage;
         
         // Compilation statistics
-        f64 m_CompilationTimeMs = 0.0;
+        f32 m_CompilationTimeMs = 0.0f;
         sizet m_SourceSizeBytes = 0;
         sizet m_CompiledSizeBytes = 0;
     };
